ci(github): run acceptance tests against Terraform 0.14.x and 0.15.x
Expanding tests them to Terraform 0.12.x and 0.13.x too would be great,
but the CheckDestroy callbacks on data source tests fail on those
versions saying the resources were not deleted, when they actually were.

For now 0.12.x and 0.13.x are excluded and not "officially" supported as
we don't run our test suite against them.