fix(virtual_machine): avoid timing related deletion failure
When deleting a running VM, we used to simply issue a stop call to the
VM, and then monitor the VM status until it changed to stopped before
moving on to purging the VM.

On the backend, stopping a VM is handled by a background asynchronous
task, and there's a very tiny window between the VM reporting a stopped
state, and the task actually finishing. If a purge VM request is sent in
that window, the purge request will fail due to being blocked by stop
task which has not been marked as done yet.

Hence we take a more active approach here, and actually monitor the task
returned by the stop request, until it is marked as done.

There is still a tiny change this error could happen if Terraform tries
to delete a VM which is already in the middle of stopping or shutting
down, as we then do not know what the Task ID is which is responsible
for the stop/shutdown request. Hence we have to fall back to simply
checking the VM state itself. Ideally we could retry the purge request
itself if it fails with a task blocked error, but error reporting from
the Katapult API client isn't not granular enough for that yet.