chore(test): rename TestTools to testTools, add Ctx key replacing Meta.Ctx
TestTools is a solely internal thing, so lets name it to indicate
that.

The Ctx field on Meta is only meant to be used by tests, so
lets get rid of it and store the context on testTools instead. The
reason it was originally added to Meta was cause sweepers need it, but
they should work just fine with context.TODO(), which is what Meta's Ctx
field was being set to anyway within tests.