feat(resource/virtual_machine): add support for setting number and size of disks during creation
Adds a repeatable "disk" block with "name" and "size" fields which are
used to define what disks should be created and attached to the VM
during creation.

Due to this not being a full disk resource implementation, but a
creation-only property of virtual machines, any changes to disk blocks
after VM creation will make Terraform want to replace the whole VM.

Fully managing disks as separate resources is planned, but out of scope
of this change.