chore(destroy): more reliably wait for trash objects to be removed
Check trash objects directly, and wait for them to disappear. Instead of
simply waiting for the purge task we triggered to complete.

Some purge operations need to run as separate background tasks in
Katapult, meaning the main task for the purge trash object request may
finish before the object itself is fully deleted.