feat(destroy): add skip_trash_object_purge provider option
Some resources in Katapult are "trashable", meaning when they are
deleted, they are instead soft-deleted and moved to the "Trash".

Terraform handles trashable resources by performing the regular delete
that moves them to trash, followed by purging the resource from the
trash. Effectively permanently deleting the resource.

Permanent deletion may not always be the desired behavior however, hence
this new `skip_trash_object_purge` option. When enabled, trashable
resources will be soft-deleted moving them to the trash, and then left
alone to be automatically purged from the trash by Katapult after 48
hours.

Resources which have any unique values, will have their unique values
suffixed with a random or object-unique value to avoid conflicts when
replacing resources with Terraform. Without this, a new replacement VM
with the same hostname for example, could not be created while the old
VM with the same hostname was still in the trash.