fix(load-balancer): handle removal of all resource IDs
The plan modifier had a bug where it would not correctly pick up full
removal of all resource IDs. This was due to the given plan still
containing the old value when it has been removed. Hence we need to
check the current config value as well to detect full removal.