fix(v6provider/refresh): refreshing missing resources now clears them from state
The new resources implemented with the Plugin Framework had a bug where
they would not clear our state of resources which no longer exist in the
API.

This has now been fixed, so resources deleted outside of Terraform are
correctly detected as missing, yielding a plan to recreate them.

Explicitly testing this behavior seems to be fraught with annoyance and
flickering results, so for now I've left that out.