fix(tags): ensure name can be changed when no color is specified
Without using state for unknown values, it would end up sending a blank
color value when changing the name, causing the API to return an error.

This also avoids the `color` attribute being marked as "known after
apply" when it's not specified in the configuration and name is being
changed.